# src/main/java/com/devweb/plocadora/web/api/LocacaoApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.devweb.plocadora.web.api;

import com.devweb.plocadora.web.model.AtualizarLocacaoApiModel;
import com.devweb.plocadora.web.model.ErroRequisicaoApiModel;
import com.devweb.plocadora.web.model.LocacaoApiModel;
import com.devweb.plocadora.web.model.LocacaoCriadaApiModel;
import com.devweb.plocadora.web.model.NovaLocacaoApiModel;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "locacao", description = "the locacao API")
public interface LocacaoApi {

    /**
     * DELETE /locacao/{locacao_id} : 
     *
     * @param locacaoId  (required)
     * @return No Content (status code 204)
     *         or Bad Request (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "deleteLocacaoById",
        summary = "",
        responses = {
            @ApiResponse(responseCode = "204", description = "No Content"),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/locacao/{locacao_id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteLocacaoById(
        @Parameter(name = "locacao_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("locacao_id") String locacaoId
    );


    /**
     * GET /locacao : Your GET endpoint
     *
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "getLocacao",
        summary = "Your GET endpoint",
        tags = {  },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LocacaoApiModel.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/locacao",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<LocacaoApiModel>> getLocacao(
        
    );


    /**
     * GET /locacao/{locacao_id} : Your GET endpoint
     *
     * @param locacaoId  (required)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "getLocacaoById",
        summary = "Your GET endpoint",
        tags = {  },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LocacaoApiModel.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/locacao/{locacao_id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<LocacaoApiModel> getLocacaoById(
        @Parameter(name = "locacao_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("locacao_id") String locacaoId
    );


    /**
     * POST /locacao : 
     *
     * @param novaLocacaoApiModel  (optional)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "postLocacao",
        summary = "",
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LocacaoCriadaApiModel.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/locacao",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<LocacaoCriadaApiModel> postLocacao(
        @Parameter(name = "NovaLocacaoApiModel", description = "") @Valid @RequestBody(required = false) NovaLocacaoApiModel novaLocacaoApiModel
    );


    /**
     * PUT /locacao/{locacao_id} : 
     *
     * @param locacaoId  (required)
     * @param atualizarLocacaoApiModel  (optional)
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "putLocacaoById",
        summary = "",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LocacaoApiModel.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroRequisicaoApiModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/locacao/{locacao_id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<LocacaoApiModel> putLocacaoById(
        @Parameter(name = "locacao_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("locacao_id") String locacaoId,
        @Parameter(name = "AtualizarLocacaoApiModel", description = "") @Valid @RequestBody(required = false) AtualizarLocacaoApiModel atualizarLocacaoApiModel
    );

}