# src/main/resources/specs/openapi.yaml
openapi: 3.1.0
...
info:
  title: plocadora
  version: '1.0'
servers:
  - url: 'http://localhost:8080'
paths:
...
  /locacao:
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Locacao'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
      operationId: get-locacao
    post:
      summary: ''
      operationId: post-locacao
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocacaoCriada'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaLocacao'
  '/locacao/{locacao_id}':
    parameters:
      - schema:
          type: string
        name: locacao_id
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Locacao'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
      operationId: get-locacao-by-id
    put:
      summary: ''
      operationId: put-locacao-by-id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Locacao'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarLocacao'
    delete:
      summary: ''
      operationId: delete-locacao-by-id
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroRequisicao'
components:
  schemas:
    ErroRequisicao:
      title: ErroRequisicao
      x-stoplight:
        id: chhmig0yg8786
      type: object
      properties:
        Status:
          type: string
        mensagem:
          type: string
...
    NovaLocacao:
      title: NovaLocacao
      x-stoplight:
        id: uo5ifu8v1l50b
      type: object
      properties:
        dt_locacao:
          type: string
          format: date
        dt_devolucao_prevista:
          type: string
          format: date
        dt_devolucao_efetiva:
          type: string
          format: date
        valor_cobrado:
          type: number
          format: double
        multa_cobrado:
          type: number
          format: double
        id_item:
          type: integer
        id_cliente:
          type: integer
    Locacao:
      title: Locacao
      x-stoplight:
        id: wjhnj7c7c6zww
      type: object
      properties:
        id:
          type: integer
        dt_locacao:
          type: string
          format: date
        dt_devolucao_prevista:
          type: string
          format: date
        dt_devolucao_efetiva:
          type: string
          format: date
        valor_cobrado:
          type: number
          format: double
        multa_cobrado:
          type: number
          format: double
        id_item:
          type: integer
        id_cliente:
          type: integer
    LocacaoCriada:
      title: LocacaoCriada
      x-stoplight:
        id: 62z5nj3okag67
      type: object
      properties:
        id:
          type: integer
        dt_locacao:
          type: string
          format: date
        dt_devolucao_prevista:
          type: string
          format: date
        dt_devolucao_efetiva:
          type: string
          format: date
        valor_cobrado:
          type: number
          format: double
        multa_cobrado:
          type: number
          format: double
        id_item:
          type: integer
        id_cliente:
          type: integer
    AtualizarLocacao:
      title: AtualizarLocacao
      x-stoplight:
        id: 1dlyince4ccjf
      type: object
      properties:
        dt_devolucao_efetiva:
          type: string
          format: date
        multa_cobrado:
          type: number
          format: double